Add GPT2 and Update Github Actions (#18)
These changes make a few large changes. It adds GPT-2 with scripts for performing end-to-end tests
with numerical validation. It also includes updates to the github actions so that we run verification that:

Unittests are passing (on pull)
Pinned versions are automagically updated (nightly)
GPT-2 is still working (nightly)
GPT-2 sample run executed below:
https://github.com/rsuderman/iree-jax/runs/6264872065?check_suite_focus=true