Fix tests and bumping scrips (#64)
The default configuration was changed to had a batch size of 1. This breaks the export tests as it
requires atleast a batch size of two. Changed to default batch size of 4 and added custom configuration
for tests that required batch size of 1.

Also includes fixes for pip install locations and tests.